minor - bug - scaler wasn't correct for history

+++ app/views.py
@@ -399,7 +399,7 @@ def history(id):
     
     return  render_template('history.html', id = id, \
         a = data['TestLog'].a, t = data['TestLog'].t, wronganswers = wronganswers, rightanswers = rightanswers, xmax = xmax, pred = pred,\
-        scaler = int(current_app.config['SAMPLE_SCALER']), \
+        scaler = float(current_app.config['SAMPLE_SCALER']), \
         curtest = curtest, cnt = cnt, \
         date = data['TestLog'].start_time, \
         avg_answered = int(current_app.config['SESSION_REDIS'].get('avg_answered') or 0), \
